Addressed draft PR comments from Zhen pt.2
- Removed more unnecessary wrapper classes
- Changed formatting of enums to camelcase
- Renamed some functions
- Changed some structs to internal/private
- Restructured example to frame all interceptor code as part of the
example, rather than the settled way of integration into gRPC
